httpbakery: fix OpenWebBrowser for Windows
