+++ httpbakery/client.go
@@ -1,7 +1,6 @@
 package httpbakery
 
 import (
-	"bytes"
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
@@ -10,8 +9,6 @@ import (
 	"net/http"
 	"net/http/cookiejar"
 	"net/url"
-	"os/exec"
-	"runtime"
 	"strings"
 
 	"github.com/juju/loggo"
@@ -565,31 +562,4 @@ func (j *cookieLogger) SetCookies(u *url.URL, cookies []*http.Cookie) {
 		logger.Debugf("\t%d. path %s; name %s", i, c.Path, c.Name)
 	}
 	j.CookieJar.SetCookies(u, cookies)
-}
-
-var browser = map[string]string{
-	"linux":   "sensible-browser",
-	"windows": "start",
-	"darwin":  "open",
-}
-
-// OpenWebBrowser opens a web browser at the
-// given URL. If the OS is not recognised, the URL
-// is just printed to standard output.
-func OpenWebBrowser(url *url.URL) error {
-	if b := browser[runtime.GOOS]; b != "" {
-		cmd := exec.Command(b, url.String())
-		data, err := cmd.CombinedOutput()
-		if err == nil {
-			return nil
-		}
-		if err != exec.ErrNotFound {
-			if _, ok := err.(*exec.ExitError); ok {
-				return errgo.Newf("cannot open web browser: %s", bytes.TrimSpace(data))
-			}
-			return errgo.Notef(err, "cannot open web browser")
-		}
-	}
-	fmt.Printf("Please visit this web page:\n%s\n", url)
-	return nil
 }
